pyproject.toml
[tool.poetry]
name = "contacts_calendar"
version = "0.5.0"
description = ""
authors = ["Patrick Vilchez <pvilchez794@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.2"
python-dotenv = "^0.19.2"
django-phone-field = "^1.8.1"
django-livereload-server = "^0.4"
psycopg2-binary = "^2.9.3"
whitenoise = "^5.3.0"
gunicorn = "^20.1.0"
dj-database-url = "^0.5.0"
django-tailwind = "^3.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
